New chat field is now being cleared after creating new chat. Moved Profile to components from pages
