+++ Stavitskii_Nikita/15/src/components/ChatList/ChatList.jsx
@@ -44,17 +44,17 @@ export class ChatList extends Component {
                     let message = chat.messages[chat.messages.length - 1];
                     const secondaryValue = message === undefined ? 'No messages in chat' : (message.author + ': ' + message.text);
                     return (
-                    <div key={index}>
-                        <Link to={chat.link}>
-                            <ListItem button className="chat-item">
-                                <ListItemText primary={chat.name} secondary={secondaryValue}/>
-                            </ListItem>
-                        </Link>
-                        <Divider variant="inset" component="li" />
-                    </div>
-                )})}
+                        <div key={index}>
+                            <Link to={chat.link}>
+                                <ListItem button className="chat-item">
+                                    <ListItemText primary={chat.name} secondary={secondaryValue}/>
+                                </ListItem>
+                            </Link>
+                            <Divider variant="inset" component="li" />
+                        </div>
+                    )})}
                 <ListItem>
-                    <TextField fullWidth name="input" label="Enter new chat name" onChange={this.handleChange} />
+                    <TextField fullWidth name="input" label="Enter new chat name" value={title} onChange={this.handleChange} />
                     <ListItemIcon>
                         {/*<AddCircleIcon onClick={this.handleAddChat} />*/}
                         <Button color="inherit" onClick={this.handleAddChat}><AddCircleIcon /></Button>
